refactor(Product) - Profanities를 infra에서 도메인 영역으로 이동
