+++ src/main/java/kitchenpos/products/infra/PurgomalumClient.java
@@ -1,5 +1,6 @@
 package kitchenpos.products.infra;
 
+import kitchenpos.products.domain.Profanities;
 import org.springframework.boot.web.client.RestTemplateBuilder;
 import org.springframework.stereotype.Component;
 import org.springframework.web.client.RestTemplate;
